getAAWithFakesFromCutoutFixtureId
