+++ shared-data/js/helpers/deckConfiguration/index.ts
@@ -1,3 +1,3 @@
+export * from './getAddressableAreaFrom'
 export * from './getFixtureFrom'
-export * from './getVisualSlotFrom'
 export * from './getVisualSlotFrom'
